+++ src/Form/Type/LexicalFormType.php
@@ -53,9 +53,11 @@ final class LexicalFormType extends AbstractType
      */
     public const AVAILABLE_BUTTONS = [
         'bold', 'italic', 'underline', 'strikethrough', 'subscript', 'superscript',
+        'align-left', 'align-center', 'align-right', 'align-justify',
         'bullet', 'number',
         'indent', 'outdent',
         'link', 'unlink',
+        'source',
     ];
 
     /**
@@ -66,11 +68,15 @@ final class LexicalFormType extends AbstractType
     public const DEFAULT_TOOLBAR = [
         'bold', 'italic', 'underline', 'strikethrough', 'subscript', 'superscript',
         self::SEPARATOR,
+        'align-left', 'align-center', 'align-right', 'align-justify',
+        self::SEPARATOR,
         'bullet', 'number',
         self::SEPARATOR,
         'indent', 'outdent',
         self::SEPARATOR,
         'link', 'unlink',
+        self::SEPARATOR,
+        'source',
     ];
 
     /**